Flatten src/, rename validation→survey, consolidate utilities - babelbetes/src/* moved to babelbetes/ root; merged other functions into pandas_helper - Fix carbs extraction bug in run_functions.py
