+++ babelbetes/drawing.py
@@ -6,10 +6,8 @@ from datetime import timedelta
 import matplotlib.pyplot as plt
 import matplotlib.dates as mdates
 import numpy as np
-import importlib 
-from babelbetes.src import pandas_helper
-importlib.reload(pandas_helper)   
-from babelbetes.src.pandas_helper import get_hour_of_day
+from babelbetes import pandas_helper
+from babelbetes.pandas_helper import get_hour_of_day
     
 colors = {'Bolus': 'red', 'Basal': 'blue', 'CGM': 'darkgray'}
