+++ babelbetes/studies/dataset_initializer.py
@@ -1,7 +1,7 @@
 
 import os
 from babelbetes.studies import StudyDataset, IOBP2, Flair, PEDAP, DCLP3, DCLP5, ReplaceBG, Loop, T1DEXI, T1DEXIP
-from babelbetes.src.logger import Logger
+from babelbetes.logger import Logger
 logger = Logger.get_logger(__name__)
 
 def initialize_datasets(data_path, subset=False) -> dict[str,StudyDataset]:
