+++ babelbetes/studies/flair.py
@@ -8,10 +8,10 @@ import numpy as np
 from functools import cached_property
 
 from babelbetes.studies.studydataset import StudyDataset
-from babelbetes.src.find_periods import find_periods
-from babelbetes.src.pandas_helper import get_df
-from babelbetes.src.date_helper import parse_flair_dates, convert_duration_to_timedelta
-from babelbetes.src import pandas_helper
+from babelbetes.pandas_helper import find_periods
+from babelbetes.pandas_helper import get_df
+from babelbetes.date_helper import parse_flair_dates, convert_duration_to_timedelta
+from babelbetes import pandas_helper
 
 
 def merge_basal_and_temp_basal(df):
